# PERF: Add DynamicThreadedGenerateData to Riesz and Wavelet Bank Generators