--- a/include/itkWaveletFrequencyFilterBankGenerator.h
+++ b/include/itkWaveletFrequencyFilterBankGenerator.h
@@ -131,7 +131,8 @@ protected:
   ~WaveletFrequencyFilterBankGenerator() override {}
   void PrintSelf(std::ostream & os, Indent indent) const override;
 
-  void GenerateData() override;
+  void BeforeThreadedGenerateData() override;
+  void DynamicThreadedGenerateData(const OutputImageRegionType & threadRegion) override;
 
 private:
   unsigned int           m_HighPassSubBands;