PERF: Add DynamicThreaded to WaveletBankGenerator
Non-scientific perf test: From 13.96s to 12.07s

PERF: Add DynamicThreaded to RieszBankGenerator
Non-scientific perf test: From 14.45s to 13.96s

Perf test was using a simple: `ctest -L "IsotropicWavelets"`