@phcerdan 2019-01-18
It is multithreading just the bank generation, not the whole pyramid.

But I am not sure, the tests are kind of toy examples. A more thorough performance test hasn't been pursued. However, an approximate ~15% of increased performance doesn't hurt either!